Create numberDigitProductSumDisplay.java
Write a program that reads one integer from the keyboard.

If the number consists of three digits, then you need to display the sum of the digits of the number. If it consists of two — the product of digits, if it's one digit — display the number.

In all other cases, do not output anything.